drf/customers/serializers.py
from rest_framework.serializers import ModelSerializer

from .models import Customer


class CustomerSerializer(ModelSerializer):
    class Meta:
        model = Customer
        fields = ('id', 'username', 'full_name', 'email')
